Add textbox/textlog state to screen

+++ api/src/main/java/nl/weeaboo/vn/render/IDrawBuffer.java
@@ -25,7 +25,7 @@ public interface IDrawBuffer {
     void drawQuad(IDrawTransform transform, int argb, ITexture tex, Area2D bounds, Area2D uv);
 
     void drawText(IDrawTransform transform, double dx, double dy, ITextLayout textLayout,
-            float visibleGlyphs);
+            double visibleGlyphs);
 
     void drawLayer(int layerId, ILayer layer);
