+++ api/src/main/java/nl/weeaboo/vn/text/ITextRenderState.java
@@ -6,28 +6,28 @@ import nl.weeaboo.styledtext.layout.ITextLayout;
 
 public interface ITextRenderState extends IMultiLineText, Serializable {
 
-    final float ALL_GLYPHS_VISIBLE = 999999;
+    final double ALL_GLYPHS_VISIBLE = 999999;
 
 	/**
 	 * Returns the tags of the TextStyle of the characters at the specified location or <code>null</code> if no hit.
 	 */
-    int[] getHitTags(float cx, float cy);
+    int[] getHitTags(double cx, double cy);
 
-    void increaseVisibleText(float textSpeed);
+    void increaseVisibleText(double textSpeed);
 
     /**
      * @see #setVisibleText(float)
      */
-    float getVisibleText();
+    double getVisibleText();
 
     /**
      * Sets the number of visible glyphs relative to the current start line.
      *
-     * @see #setVisibleText(int, float)
+     * @see #setVisibleText(int, double)
      */
-    void setVisibleText(float visibleGlyphs);
+    void setVisibleText(double visibleGlyphs);
 
-    void setVisibleText(int startLine, float visibleGlyphs);
+    void setVisibleText(int startLine, double visibleGlyphs);
 
     ITextLayout getVisibleLayout();
